requirements.txt
asgiref==3.2.10
Django==3.0.5
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.1
djongo==1.3.3
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.2.4
